requirements.txt
psycopg2-binary>=2.9.0
pyinstaller>=6.0.0
